refactor(http): move endpoint error to http package
